+++ client/src/features/common/ImageThumbnail.tsx
@@ -9,6 +9,7 @@ import { IImages } from '../../types/IImages'
 import StyledImageThumbnail from '../../styles/features/common/ImageThumbnail'
 import theme from '../../styles/theme'
 import { stripYaleIdPrefix } from '../../lib/parse/data/helper'
+import { pushSiteImproveEvent } from '../../lib/siteImprove'
 
 interface IProps {
   imageInfo: IImages
@@ -39,6 +40,13 @@ const ImageThumbnail: React.FC<IProps> = ({ imageInfo, linkUrl }) => {
       {linkUrl !== undefined ? (
         <Link
           to={`/view/${stripYaleIdPrefix(linkUrl)}`}
+          onClick={() =>
+            pushSiteImproveEvent(
+              'Entity Link',
+              'Selected',
+              'Results Snippet Link',
+            )
+          }
           data-testid="image-link"
         >
           <img