+++ client/src/features/error/ErrorPage.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useEffect } from 'react'
 import Col from 'react-bootstrap/Col'
 import Row from 'react-bootstrap/Row'
 
@@ -6,6 +6,7 @@ import { BorderedDiv } from '../../styles/shared/BorderedDiv'
 import StyledErrorPage from '../../styles/features/error/ErrorPage'
 import theme from '../../styles/theme'
 import SearchContainer from '../search/SearchContainer'
+import { pushSiteImproveEvent } from '../../lib/siteImprove'
 
 interface IProps {
   code: number | undefined
@@ -20,32 +21,38 @@ const getMessage = (code: number | undefined): string => {
   }
 }
 
-const ErrorPage: React.FC<IProps> = ({ code }) => (
-  <StyledErrorPage className="row mx-0">
-    <BorderedDiv className="col section">
-      <Row>
-        <Col xs={12}>
-          <div className="message d-flex justify-content-md-center">
-            <div>{getMessage(code)}</div>
-          </div>
-          <div className="search">
-            <SearchContainer
-              className="errorPageSearchBox"
-              bgColor={theme.color.white}
-              id="error-page-search-box"
-            />
-          </div>
-        </Col>
-        {/* TODO: add image
-        <Col xs={12} md={4}">
-          <div className="image-container">
-            <img src="" alt="" />
-          </div>
-        </Col>
-        */}
-      </Row>
-    </BorderedDiv>
-  </StyledErrorPage>
-)
+const ErrorPage: React.FC<IProps> = ({ code }) => {
+  useEffect(() => {
+    pushSiteImproveEvent('Error', 'Triggered', 'Page Not Found 404')
+  }, [])
+
+  return (
+    <StyledErrorPage className="row mx-0">
+      <BorderedDiv className="col section">
+        <Row>
+          <Col xs={12}>
+            <div className="message d-flex justify-content-md-center">
+              <div>{getMessage(code)}</div>
+            </div>
+            <div className="search">
+              <SearchContainer
+                className="errorPageSearchBox"
+                bgColor={theme.color.white}
+                id="error-page-search-box"
+              />
+            </div>
+          </Col>
+          {/* TODO: add image
+          <Col xs={12} md={4}">
+            <div className="image-container">
+              <img src="" alt="" />
+            </div>
+          </Col>
+          */}
+        </Row>
+      </BorderedDiv>
+    </StyledErrorPage>
+  )
+}
 
 export default ErrorPage